change: 노래 추천 버튼을 화면 크기별 위치로 분리
lg 미만에서는 챗봇처럼 전역 플로팅 버튼으로 띄워 신청 리스트 가림을 막고, lg 이상에서는 기존처럼 섹션 내 버튼으로 표시한다. 버튼·모달·추천 신청 로직을 공용 컴포넌트와 훅으로 추출해 두 위치에서 재사용한다.
